perf: lazy load react-joyride to improve initial bundle size
Move react-joyride (~85KB) to a separate lazy-loaded chunk that's
only downloaded when the user actually starts the tour. This reduces
the main bundle by ~85KB and should improve Lighthouse performance score.
