+++ src/lib/onboarding-tour/tour-context.tsx
@@ -1,19 +1,14 @@
-import { useState, useCallback, useEffect, useMemo } from 'react';
-import Joyride, { ACTIONS, EVENTS, STATUS } from 'react-joyride';
-import type { TourContextValue, TourProviderProps, TourCallbackData, TourStep } from './types';
+import { useState, useCallback, useEffect, useMemo, lazy, Suspense } from 'react';
+import type { TourContextValue, TourProviderProps } from './types';
 import { TourContext } from './tour-context-value';
 import { DEFAULT_TOUR_STEPS } from './tour-steps';
-import { tourStyles, tourLocale, floaterProps } from './tour-styles';
-import {
-  getTourState,
-  markTourCompleted,
-  markTourDismissed,
-  markStepViewed,
-  incrementStartCount,
-  resetTourState,
-  shouldShowTour,
-} from './tour-storage';
+import { getTourState, incrementStartCount, resetTourState, shouldShowTour } from './tour-storage';
 import { trackEvent } from '@/lib/posthog-lazy';
+
+// Lazy load the entire Joyride wrapper to avoid impacting initial page load (~40KB savings)
+const LazyJoyrideWrapper = lazy(() =>
+  import('./lazy-joyride-wrapper').then((mod) => ({ default: mod.LazyJoyrideWrapper }))
+);
 
 /**
  * Provider component for the onboarding tour
@@ -94,80 +89,21 @@ export function TourProvider({
     trackEvent('onboarding_tour_reset');
   }, []);
 
-  // Handle Joyride callbacks
-  const handleJoyrideCallback = useCallback(
-    (data: TourCallbackData) => {
-      const { action, index, status, type } = data;
-      const currentStep = steps[index] as TourStep | undefined;
+  // Callbacks for the lazy-loaded Joyride wrapper
+  const handleStepChange = useCallback((index: number) => {
+    setStepIndex(index);
+  }, []);
 
-      // Track step views
-      if (type === EVENTS.STEP_AFTER && currentStep) {
-        markStepViewed(currentStep.id);
-        trackEvent('onboarding_tour_step_completed', {
-          step_id: currentStep.id,
-          step_name: currentStep.id.replace(/-/g, ' '),
-          step_number: index + 1,
-          total_steps: steps.length,
-          progress_percentage: Math.round(((index + 1) / steps.length) * 100),
-          category: currentStep.category,
-        });
-      }
+  const handleTourComplete = useCallback(() => {
+    setIsRunning(false);
+    setTourState(getTourState());
+    onComplete?.();
+  }, [onComplete]);
 
-      // Handle tour completion
-      if (status === STATUS.FINISHED) {
-        setIsRunning(false);
-        markTourCompleted();
-        setTourState(getTourState());
-        trackEvent('onboarding_tour_completed', {
-          total_steps: steps.length,
-          steps_completed: steps.length,
-          completion_rate: 100,
-        });
-        // Call the completion callback to handle navigation
-        onComplete?.();
-        return;
-      }
-
-      // Handle tour skip/close (abandonment)
-      if (status === STATUS.SKIPPED || action === ACTIONS.CLOSE) {
-        setIsRunning(false);
-        markTourDismissed();
-        setTourState(getTourState());
-        const stepsCompleted = index;
-        const completionRate = Math.round((stepsCompleted / steps.length) * 100);
-        trackEvent('onboarding_tour_abandoned', {
-          total_steps: steps.length,
-          steps_completed: stepsCompleted,
-          completion_rate: completionRate,
-          abandoned_at_step: index,
-          abandoned_step_id: currentStep?.id,
-          abandoned_step_name: currentStep?.id?.replace(/-/g, ' '),
-          action: action === ACTIONS.CLOSE ? 'closed' : 'skipped',
-        });
-        return;
-      }
-
-      // Handle navigation
-      if (type === EVENTS.STEP_AFTER) {
-        if (action === ACTIONS.NEXT) {
-          setStepIndex(index + 1);
-        } else if (action === ACTIONS.PREV) {
-          setStepIndex(index - 1);
-        }
-      }
-
-      // Handle target not found
-      if (type === EVENTS.TARGET_NOT_FOUND) {
-        // Skip to next step if target not found
-        setStepIndex(index + 1);
-        trackEvent('onboarding_tour_target_not_found', {
-          step_id: currentStep?.id,
-          step_index: index,
-        });
-      }
-    },
-    [steps, onComplete]
-  );
+  const handleTourStop = useCallback(() => {
+    setIsRunning(false);
+    setTourState(getTourState());
+  }, []);
 
   const contextValue = useMemo<TourContextValue>(
     () => ({
@@ -201,23 +137,19 @@ export function TourProvider({
   return (
     <TourContext.Provider value={contextValue}>
       {children}
-      <Joyride
-        callback={handleJoyrideCallback}
-        continuous
-        hideCloseButton={false}
-        run={isRunning}
-        scrollToFirstStep
-        showProgress
-        showSkipButton
-        stepIndex={stepIndex}
-        steps={steps}
-        styles={tourStyles}
-        locale={tourLocale}
-        floaterProps={floaterProps}
-        disableOverlayClose={false}
-        spotlightClicks={true}
-        disableScrolling={false}
-      />
+      {/* Only load Joyride when tour is running to avoid impacting initial bundle */}
+      {isRunning && (
+        <Suspense fallback={null}>
+          <LazyJoyrideWrapper
+            isRunning={isRunning}
+            stepIndex={stepIndex}
+            steps={steps}
+            onStepChange={handleStepChange}
+            onComplete={handleTourComplete}
+            onStop={handleTourStop}
+          />
+        </Suspense>
+      )}
     </TourContext.Provider>
   );
 }
